--- a/psi4/src/core.cc
+++ b/psi4/src/core.cc
@@ -69,15 +69,16 @@ using namespace psi;
 // Python helper wrappers
 void export_benchmarks(py::module&);
 void export_blas_lapack(py::module&);
+void export_cubeprop(py::module&);
+void export_efp(py::module&);
+void export_fock(py::module&);
+void export_functional(py::module&);
+void export_mints(py::module&);
+void export_misc(py::module&);
+void export_oeprop(py::module&);
 void export_plugins(py::module&);
 void export_psio(py::module&);
-void export_mints(py::module&);
-void export_functional(py::module&);
-void export_fock(py::module&);
-void export_oeprop(py::module&);
-void export_efp(py::module&);
-void export_cubeprop(py::module&);
-void export_misc(py::module&);
+void export_wavefunction(py::module&);
 
 // In export_plugins.cc
 void py_psi_plugin_close_all();
@@ -1486,6 +1487,7 @@ PYBIND11_PLUGIN(core) {
     export_functional(core);
     export_misc(core);
     export_fock(core);
+    export_wavefunction(core);
 
     // ??
     //py::class_<Process::Environment>(core, "Environment")