--- a/psi4/src/export_wavefunction.cc
+++ b/psi4/src/export_wavefunction.cc
@@ -25,6 +25,7 @@
  * @END LICENSE
  */
 
+#include "psi4/libmints/basisset.h"
 #include "psi4/libmints/sobasis.h"
 #include "psi4/libmints/molecule.h"
 #include "psi4/libmints/oeprop.h"