--- a/psi4/src/psi4/dfmp2/corr_grad.h
+++ b/psi4/src/psi4/dfmp2/corr_grad.h
@@ -48,8 +48,8 @@ protected:
     int bench_;
     /// Memory available, in doubles, defaults to 256 MB (32 M doubles)
     unsigned long int memory_;
-    /// Number of OpenMP threads (defaults to 1 in no OpenMP, omp_get_max_thread() otherwise)
-    int omp_num_threads_;
+    /// Number of OpenMP threads (defaults to 1 in no OpenMP, Process::environment.get_n_threads() otherwise)
+    int nhreads_;
     /// Integral cutoff (defaults to 0.0)
     double cutoff_;