@andysim 2017-03-03
Yeah, the norm thing is bad.  According to my QtCreator, it's not used anywhere so should be a trivial fix.